+++ python/tvm/relax/__init__.py
@@ -23,6 +23,7 @@ from . import ir_builder
 from . import op
 from . import parser
 from . import analysis
+from . import transform
 
 
 # Expr
